AUDIT-15 fixes: close VarInt, corpus figure, packet_ids line, lp_vec3 pins, docs
Serverbound close reads CONTAINER_ID VarInt (was u8+fail-closed);
sweep corpus 6273->6258 actual; explicit serverbound-11 id check;
LpVec3 canonical-form unit tests; cow speed exact pin; loot doc.

+++ crates/server/src/game/session.rs
@@ -768,7 +768,7 @@ impl Game {
                     return Ok(());
                 }
                 if let Some(expected) = current_window
-                    && window_id != expected
+                    && window_id != i32::from(expected)
                 {
                     debug!(id = %id, window_id, "ignoring a close for a stale window");
                     return Ok(());
